requirement.txt
asgiref==3.5.2
Django==3.2.10
djangorestframework==3.12.2
pytz==2022.2.1
sqlparse==0.4.2
